feat!: try out ESM only
BREAKING CHANGES: remove CJS and IIFE
