+++ .github/workflows/cd.yml
@@ -56,10 +56,8 @@ jobs:
       - run: npm ci
       - run: npm run lint
       - run: npm run test:coverage
-      - run: npm run build:lib
+      - run: npm run build
       - run: npm run build:demo
-      - run: npm run build:dist
-      - run: npm run build:standalone
       - run: npx --yes wet-run@1.0.1 release ${{ inputs.version }} ${{ inputs.dryrun && '--dry-run' || '' }} ${{ inputs.prerelease && format('--prerelease {0}', inputs.prerelease) || '' }} --provenance --github-release --verbose
       - name: Get NPM version
         id: npm-version
