+++ .github/workflows/ci.yml
@@ -58,10 +58,8 @@ jobs:
         node-version: ${{ matrix.node-version }}
         cache: npm
     - run: npm ci
-    - run: npm run build:lib
+    - run: npm run build
     - run: npm run build:demo
-    - run: npm run build:dist
-    - run: npm run build:standalone
 
   deploy-preview:
     # Grant GITHUB_TOKEN the permissions required to make a Pages deployment
@@ -82,7 +80,7 @@ jobs:
         with:
           node-version: 20
       - run: npm ci
-      - run: npm run build:lib
+      - run: npm run build
       - run: npm run build:demo
       - uses: actions/upload-pages-artifact@v3
         with:
