+++ .gitignore
@@ -10,3 +10,4 @@ yarn-error.log
 .idea/
 .vscode/
 /disttest/
+/dist/
