--- .npmignore
@@ -1,12 +1,0 @@
-.github
-babel.config.js
-codecov.yml
-coverage
-src
-test
-webpack/*
-yarn-error.log
-/demo/
-/examples/
-/disttest/
-/scripts/
