+++ types/lazy/index.d.ts
@@ -1,2 +1,2 @@
-export * from '../lib'
-export { default } from '../lib'
+export * from '../'
+export { default } from '../'
